fix merge problems
